requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
discord==2.1.0
discord.py==2.1.0
distlib==0.3.5
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
environ==1.0
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
multidict==6.0.4
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
platformdirs==2.5.2
psycopg2-binary==2.9.5
pycodestyle==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.12
virtualenv==20.16.3
yarl==1.8.2
